extracted argument_interface and nargs
